use of setup_project fonction for all the projects currently implemented
